This fixes this compile warning: util.c:4052: warning: passing argument 2 of ‘process_chunked_data’ from incompatible pointer type util.c:3861: note: expected ‘gsize *’ but argument is of type ‘long unsigned int *’
Someone could do with checking our use of various different types (gsize,
gssize, unsigned long, I forget what else) throughout the url parsing and
retrieving things.

--HG--
extra : convert_revision : bd005861cf3abdbc22c2cbbc9e207663d787b069

+++ libpurple/util.c
@@ -65,7 +65,7 @@ struct _PurpleUtilFetchUrlData
 	gboolean got_headers;
 	gboolean has_explicit_data_len;
 	char *webdata;
-	unsigned long len;
+	gsize len;
 	unsigned long data_len;
 	gssize max_len;
 	gboolean chunked;
